[Refactor] 로또 출력 데이터 관련 유틸 분리
